Refatorado data_analysis.py Concluido refatoração de run.py
